Roll Update
No the roll  while ending it will make the player stay in a place for a small period of time. For the palyers to not spam it all the time
