+++ Project Smash/Assets/_Scripts/Attributes/ArmoredEnemy.cs.meta
